fix: entity engine

+++ core/utils/keyboard-utility.go
@@ -6,7 +6,7 @@ import (
 	rl "github.com/gen2brain/raylib-go/raylib"
 )
 
-func GetVecForKeyboard(v int) (res base.Vec[float32]) {
+func GetVecForKeyboard(v float32) (res base.Vec[float32]) {
 	unit := float32(v) * rl.GetFrameTime()
 	if rl.IsKeyDown(rl.KeyD) {
 		res.AddScalars(unit, 0)
